fix email notification settings for android
